While working on #946 I realized it takes a lot of time to start the test suite *(which is a bit frustrating when doing TDD and having to run the certain tests over and over)* 🐌 

So I added Spotify's [bootsnap](https://github.com/Shopify/bootsnap). It's a drop in solution for basically a faster start of the app (and does not conflict with spring) and will be part of the default Gemfile in the upcoming Rails version.

Small proof of concept 😉 :
```shell
# on master
~/code/rgsoc/rgsoc-teams master
❯ time rake environment
rake environment  2.30s user 1.08s system 96% cpu 3.506 total

# on this branch
~/code/rgsoc/rgsoc-teams add-bootsnap*
❯ time rake environment
rake environment  1.19s user 0.33s system 92% cpu 1.634 total
```